client/requirements.txt
requests
ptpython
prompt_toolkit
websocket-client
pandas
matplotlib
numpy
scipy